Update codes to handle different ego networks separately
All the codes have been changed so that instead of processing on all of the network at once, the codes now deal with each ego network (cluster) separately. This was done so that Hadamard vectors can be generated and used easily, without needing to compute absolutely enormous amount of data. This in no way should affect the quality of the results as individually, ego networks are quite sparse from each other.